+++ themes/theme-docs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @marigold/theme-docs
 
+## 3.0.3
+
+### Patch Changes
+
+- 0e8211b: chore([DST-853]): Refa styles for `<Menu>` button
+- Updated dependencies [0e8211b]
+- Updated dependencies [af401e5]
+- Updated dependencies [534ad77]
+  - @marigold/components@12.0.1
+  - @marigold/system@12.0.1
+
 ## 3.0.2
 
 ### Patch Changes